kali/requirements.txt
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3
gunicorn==20.1.0
Django==3.2.16
gevent==22.10.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
Pillow=10.3.0